requirements.txt
click==7.0
pyserial==3.4
tabulate==0.8.7